backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for Manobela"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "mediapipe>=0.10.31",
    "opencv-python>=4.12.0.88",
    "pydantic-extra-types>=2.11.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
